requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
boto3==1.26.154
botocore==1.29.154
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cors-headers==4.0.0
django-phonenumber-field==7.1.0
django-rest-swagger==2.2.0
django-simple-history==3.3.0
django-smtp-ssl==1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.13
Pillow==9.5.0
pipreqs==0.4.13
psycopg2-binary
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
s3transfer==0.6.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
yarg==0.1.9
